refactor: write prompt using pigment
