+++ packages/create-react-native-library/package.json
@@ -48,10 +48,8 @@
     "github-username": "^6.0.0",
     "kleur": "^4.1.4",
     "ora": "^5.4.1",
-    "pigment": "file:.yalc/pigment",
-    "prompts": "^2.4.2",
-    "validate-npm-package-name": "^4.0.0",
-    "yargs": "^17.5.1"
+    "pigment": "^0.3.10",
+    "validate-npm-package-name": "^4.0.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.24.8",
@@ -63,8 +61,6 @@
     "@types/dedent": "^0.7.0",
     "@types/ejs": "^3.1.1",
     "@types/fs-extra": "^9.0.13",
-    "@types/prompts": "^2.0.14",
-    "@types/validate-npm-package-name": "^3.0.3",
-    "@types/yargs": "^17.0.10"
+    "@types/validate-npm-package-name": "^3.0.3"
   }
 }
