+++ packages/react-native-builder-bob/vitest.config.mjs
@@ -2,6 +2,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    exclude: ['lib/'],
+    exclude: ['**/lib/**', '**/node_modules/**'],
   },
 });
